requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2  
google-auth==2.22.0        
pinecone-client==3.0.1     
requests==2.31.0  
flask-cors         